moved useDiagnostics to hooks dir
